requirements.txt
torch
opencv-python
easyocr
numpy
pillow
torchvision
ultralytics
paho-mqtt
